whisper-srt/requirements.txt
flask
openai-whisper
gunicorn
